Bug 1251541: Pass arguments to AVRCP passthrough command as uint8_t, r=shuang
The AVRCP passthough-command notification doesn't require |int| type
arguments. Using |uint8_t| is sufficient. This change is required to
prepare replacing |nsAutoArrayPtr<>| with |UniquePtr<[]>|.

+++ dom/bluetooth/bluedroid/BluetoothAvrcpManager.h
@@ -108,7 +108,7 @@ private:
 
   void VolumeChangeNotification(uint8_t aVolume, uint8_t aCType) override;
 
-  void PassthroughCmdNotification(int aId, int aKeyState) override;
+  void PassthroughCmdNotification(uint8_t aId, uint8_t aKeyState) override;
 
   BluetoothAddress mDeviceAddress;
   RefPtr<BluetoothProfileController> mController;
